+++ labo1/MatrixReloaded/operators/Multiply.hpp
@@ -11,7 +11,7 @@
  */
 class Multiply : public Operator {
 public:
-	int apply(int a, int b) const override;
+	long apply(long a, long b) const override;
 };
 
 #endif // MATRIXRELOADED_MULTIPLY_HPP
