+++ labo1/MatrixReloaded/operators/Subtract.cpp
@@ -1,9 +1,5 @@
-//
-// Created by Valentin Kaelin on 03.03.22.
-//
-
 #include "Subtract.hpp"
 
-int Subtract::apply(int a, int b) const {
+long int Subtract::apply(long a, long b) const {
 	return a - b;
 }
